+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - chore: remove what has remained from poetry in the codebase
 - refactor(core): avoid truncating or coloring logs in log files
 - feat(web-ui): add web-ui service
+- feat(web-ui): process input demands, dispatched on the bus
 
 ## Version 1.0.0